requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
python-decouple==3.4
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
